ref(scanner): Extract smi5879-gate-check watched-path constants
SMI-6033 Wave 4: two housekeeping changes needed to wire the two new
Wave 4 detector-pairs (decoy_misdirection, the CO_SIGNAL_MIN_SEVERITY
escalation model) into this repo's existing test-closure infrastructure.

smi5879-gate-check.closure.ts crossed the 500-line file gate once Wave 2/3/4
each appended their own detector entries to CLOSURE_TEST_FILES /
CLOSURE_WATCHED_SOURCE_PATHS. Extracted those two constants verbatim into a
new sibling smi5879-gate-check.watched-paths.ts (pure data move, re-exported
so every existing import path keeps working).

scripts/tests/indexer/parity.test.ts's core<->edge equality test for the
old CODE_EXECUTION_CO_OCCURRENCE Set is rewritten for the new
CO_SIGNAL_MIN_SEVERITY map -- still an EQUALITY assertion over both keys AND
minimum-severity values, not a superset check, per the plan's explicit
instruction: an edge-only entry or a lower minimum on one side would
silently widen the prod escalation gate beyond what core does.

Refs SMI-6033
